requirements.txt
pytest
pytest-playwright
allure-pytest
playwright
flake8